feat(neoforge): raise sequence-array batch limit and rewrite the well API docs
- Raise the Molecular Sequence Rewrite Array and Assembler Matrix Sequence
  Rewrite Core logical batch limit from Integer.MAX_VALUE to Long.MAX_VALUE;
  multiplication and buffer overflow are rejected without consuming inputs.
- Add the controller's pattern-only item interface for ME storage buses and
  item automation, sync pattern edits per crystal, and keep explicit pattern
  ownership on recovery and dismantling.
- Rewrite docs/matter-research-api.md for 2.0.3 in English, add a cross-linked
  Chinese version (docs/matter-research-api.zh-CN.md), and update the API
  index, both READMEs, the batch-provider cross-links and the in-game guide.
- Refresh CHANGELOG, 2.0.3 release notes and the CurseForge description with
  the new validation results.

+++ src/main/java/com/atir/molecularmanipulator/blockentity/MolecularManipulatorBlockEntity.java
@@ -44,7 +44,7 @@ public final class MolecularManipulatorBlockEntity extends PatternProviderBlockE
         implements SegmentedPatternContainerHost {
     public static final int PATTERN_SLOTS = 360;
     public static final int PATTERNS_PER_PAGE = 36;
-    public static final long VIRTUAL_PARALLEL_LIMIT = Integer.MAX_VALUE;
+    public static final long VIRTUAL_PARALLEL_LIMIT = Long.MAX_VALUE;
     private static final int MAX_BUFFERED_TYPES = 256;
     private static final String OUTPUT_BUFFER_TAG = "output_buffer";
     private static final String OUTPUT_READY_TICK_TAG = "output_ready_tick";
